+++ models/contacts.json
@@ -1,11 +1,5 @@
 [
   {
-    "id": "qdggE76Jtbfd9eWJHrssH",
-    "name": "Chaim Lewis",
-    "email": "dui.in@egetlacus.ca",
-    "phone": "(294) 840-6685"
-  },
-  {
     "id": "drsAJ4SHPYqZeG-83QTVW",
     "name": "Kennedy Lane",
     "email": "mattis.Cras@nonenimMauris.net",
@@ -58,5 +52,11 @@
     "name": "Paul",
     "email": "Paul@gmail.com",
     "phone": "(067) 327-2007"
+  },
+  {
+    "id": "Df94PvvorbAxZygwgwkhh",
+    "name": "Ibragim",
+    "email": "Ibragim@gmail.com",
+    "phone": "(063) 337-2008"
   }
 ]
