+++ routes/api/contacts-routes.js
@@ -1,90 +1,17 @@
 const express = require('express');
-const Joi = require('joi');
 
-const contatctsServices = require('../../models/contacts');
-
-const { HttpError } = require('../../helpers');
+const contactController = require('../../controllers/contatcts-controller');
 
 const router = express.Router();
 
-const phoneRegex = /[\(\d{3}\)]? \d{3}-\d{4}/;
+router.get('/', contactController.getAllContacts);
 
-const addSchema = Joi.object({
-  name: Joi.string().min(3).max(20).required(),
-  email: Joi.string()
-    .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net'] } })
-    .required(),
-  phone: Joi.string().pattern(phoneRegex).required(),
-});
+router.get('/:contactId', contactController.getContatctById);
 
-router.get('/', async (req, res, next) => {
-  try {
-    const result = await contatctsServices.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post('/', contactController.addContatct);
 
-router.get('/:contactId', async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contatctsServices.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete('/:contactId', contactController.deleteContatctById);
 
-router.post('/', async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    console.log(error);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-
-    const result = await contatctsServices.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.delete('/:contactId', async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contatctsServices.removeContact();
-    if (!result) {
-      throw HttpError(404, `Contact with ${contactId} not found`);
-    }
-
-    res.json({
-      message: 'contact deleted',
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put('/:contactId', async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { contactId } = req.params;
-    const result = await contatctsServices.updateContact(contactId, req.body);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put('/:contactId', contactController.updateContatctById);
 
 module.exports = router;
